@@ -3,15 +3,19 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-var gulp = require('gulp');
-var path = require('path');
-var util = require('./lib/util');
-var common = require('./lib/optimize');
-var es = require('event-stream');
-var File = require('vinyl');
+const gulp = require('gulp');
+const path = require('path');
+const util = require('./lib/util');
+const common = require('./lib/optimize');
+const es = require('event-stream');
+const File = require('vinyl');
+const i18n = require('./lib/i18n');
+const standalone = require('./lib/standalone');
+const cp = require('child_process');
 
 var root = path.dirname(__dirname);
 var sha1 = util.getVersion(root);
+// @ts-ignore Microsoft/TypeScript#21262 complains about a require of a JSON file
 var semver = require('./monaco/package.json').version;
 var headerVersion = semver + '(' + sha1 + ')';
 
@@ -21,14 +25,14 @@ var editorEntryPoints = [
 	{
 		name: 'vs/editor/editor.main',
 		include: [],
-		exclude: [ 'vs/css', 'vs/nls' ],
-		prepend: [ 'out-build/vs/css.js', 'out-build/vs/nls.js' ],
+		exclude: ['vs/css', 'vs/nls'],
+		prepend: ['out-build/vs/css.js', 'out-build/vs/nls.js'],
 	},
 	{
 		name: 'vs/base/common/worker/simpleWorker',
-		include: [ 'vs/editor/common/services/editorSimpleWorker' ],
-		prepend: [ 'vs/loader.js' ],
-		append: [ 'vs/base/worker/workerMain' ],
+		include: ['vs/editor/common/services/editorSimpleWorker'],
+		prepend: ['vs/loader.js'],
+		append: ['vs/base/worker/workerMain'],
 		dest: 'vs/base/worker/workerMain.js'
 	}
 ];
@@ -70,6 +74,8 @@ function editorLoaderConfig() {
 	return result;
 }
 
+const languages = i18n.defaultLanguages.concat([]);  // i18n.defaultLanguages.concat(process.env.VSCODE_QUALITY !== 'stable' ? i18n.extraLanguages : []);
+
 gulp.task('clean-optimized-editor', util.rimraf('out-editor'));
 gulp.task('optimize-editor', ['clean-optimized-editor', 'compile-client-build'], common.optimizeTask({
 	entryPoints: editorEntryPoints,
@@ -79,14 +85,37 @@ gulp.task('optimize-editor', ['clean-optimized-editor', 'compile-client-build'],
 	bundleLoader: false,
 	header: BUNDLED_FILE_HEADER,
 	bundleInfo: true,
-	out: 'out-editor'
+	out: 'out-editor',
+	languages: languages
 }));
 
 gulp.task('clean-minified-editor', util.rimraf('out-editor-min'));
 gulp.task('minify-editor', ['clean-minified-editor', 'optimize-editor'], common.minifyTask('out-editor'));
 
+gulp.task('clean-editor-esm', util.rimraf('out-editor-esm'));
+gulp.task('extract-editor-esm', ['clean-editor-esm', 'clean-editor-distro'], function() {
+	standalone.createESMSourcesAndResources({
+		entryPoints: [
+			'vs/editor/editor.main',
+			'vs/editor/editor.worker'
+		],
+		outFolder: './out-editor-esm/src',
+		outResourcesFolder: './out-monaco-editor-core/esm',
+		redirects: {
+			'vs/base/browser/ui/octiconLabel/octiconLabel': 'vs/base/browser/ui/octiconLabel/octiconLabel.mock',
+			'vs/nls': 'vs/nls.mock',
+		}
+	});
+});
+gulp.task('compile-editor-esm', ['extract-editor-esm', 'clean-editor-distro'], function() {
+	const result = cp.spawnSync(`node`, [`../node_modules/.bin/tsc`], {
+		cwd: path.join(__dirname, '../out-editor-esm')
+	});
+	console.log(result.stdout.toString());
+});
+
 gulp.task('clean-editor-distro', util.rimraf('out-monaco-editor-core'));
-gulp.task('editor-distro', ['clean-editor-distro', 'minify-editor', 'optimize-editor'], function() {
+gulp.task('editor-distro', ['clean-editor-distro', 'compile-editor-esm', 'minify-editor', 'optimize-editor'], function () {
 	return es.merge(
 		// other assets
 		es.merge(
@@ -97,17 +126,17 @@ gulp.task('editor-distro', ['clean-editor-distro', 'minify-editor', 'optimize-ed
 
 		// package.json
 		gulp.src('build/monaco/package.json')
-			.pipe(es.through(function(data) {
+			.pipe(es.through(function (data) {
 				var json = JSON.parse(data.contents.toString());
 				json.private = false;
-				data.contents = new Buffer(JSON.stringify(json, null, '  '));
+				data.contents = Buffer.from(JSON.stringify(json, null, '  '));
 				this.emit('data', data);
 			}))
 			.pipe(gulp.dest('out-monaco-editor-core')),
 
 		// README.md
 		gulp.src('build/monaco/README-npm.md')
-			.pipe(es.through(function(data) {
+			.pipe(es.through(function (data) {
 				this.emit('data', new File({
 					path: data.path.replace(/README-npm\.md/, 'README.md'),
 					base: data.base,
@@ -124,10 +153,10 @@ gulp.task('editor-distro', ['clean-editor-distro', 'minify-editor', 'optimize-ed
 		// min folder
 		es.merge(
 			gulp.src('out-editor-min/**/*')
-		).pipe(filterStream(function(path) {
+		).pipe(filterStream(function (path) {
 			// no map files
 			return !/(\.js\.map$)|(nls\.metadata\.json$)|(bundleInfo\.json$)/.test(path);
-		})).pipe(es.through(function(data) {
+		})).pipe(es.through(function (data) {
 			// tweak the sourceMappingURL
 			if (!/\.js$/.test(data.path)) {
 				this.emit('data', data);
@@ -140,49 +169,50 @@ gulp.task('editor-distro', ['clean-editor-distro', 'minify-editor', 'optimize-ed
 			var newStr = '//# sourceMappingURL=' + relativePathToMap.replace(/\\/g, '/');
 			strContents = strContents.replace(/\/\/\# sourceMappingURL=[^ ]+$/, newStr);
 
-			data.contents = new Buffer(strContents);
+			data.contents = Buffer.from(strContents);
 			this.emit('data', data);
 		})).pipe(gulp.dest('out-monaco-editor-core/min')),
 
 		// min-maps folder
 		es.merge(
 			gulp.src('out-editor-min/**/*')
-		).pipe(filterStream(function(path) {
+		).pipe(filterStream(function (path) {
 			// no map files
 			return /\.js\.map$/.test(path);
 		})).pipe(gulp.dest('out-monaco-editor-core/min-maps'))
 	);
 });
 
-gulp.task('analyze-editor-distro', function() {
+gulp.task('analyze-editor-distro', function () {
+	// @ts-ignore Microsoft/TypeScript#21262 complains about a require of a JSON file
 	var bundleInfo = require('../out-editor/bundleInfo.json');
 	var graph = bundleInfo.graph;
 	var bundles = bundleInfo.bundles;
 
 	var inverseGraph = {};
-	Object.keys(graph).forEach(function(module) {
+	Object.keys(graph).forEach(function (module) {
 		var dependencies = graph[module];
-		dependencies.forEach(function(dep) {
+		dependencies.forEach(function (dep) {
 			inverseGraph[dep] = inverseGraph[dep] || [];
 			inverseGraph[dep].push(module);
 		});
 	});
 
 	var detailed = {};
-	Object.keys(bundles).forEach(function(entryPoint) {
+	Object.keys(bundles).forEach(function (entryPoint) {
 		var included = bundles[entryPoint];
 		var includedMap = {};
-		included.forEach(function(included) {
+		included.forEach(function (included) {
 			includedMap[included] = true;
 		});
 
 		var explanation = [];
-		included.map(function(included) {
+		included.map(function (included) {
 			if (included.indexOf('!') >= 0) {
 				return;
 			}
 
-			var reason = (inverseGraph[included]||[]).filter(function(mod) {
+			var reason = (inverseGraph[included] || []).filter(function (mod) {
 				return !!includedMap[mod];
 			});
 			explanation.push({
@@ -198,7 +228,7 @@ gulp.task('analyze-editor-distro', function() {
 });
 
 function filterStream(testFunc) {
-	return es.through(function(data) {
+	return es.through(function (data) {
 		if (!testFunc(data.relative)) {
 			return;
 		}
