@@ -27,15 +27,19 @@ const common = require('./lib/optimize');
 const nlsDev = require('vscode-nls-dev');
 const root = path.dirname(__dirname);
 const commit = util.getVersion(root);
+// @ts-ignore Microsoft/TypeScript#21262 complains about a require of a JSON file
 const packageJson = require('../package.json');
+// @ts-ignore Microsoft/TypeScript#21262 complains about a require of a JSON file
 const product = require('../product.json');
 const crypto = require('crypto');
 const i18n = require('./lib/i18n');
 const glob = require('glob');
 const deps = require('./dependencies');
 const getElectronVersion = require('./lib/electron').getElectronVersion;
+const createAsar = require('./lib/asar').createAsar;
 
 const productionDependencies = deps.getProductionDependencies(path.dirname(__dirname));
+// @ts-ignore
 const baseModules = Object.keys(process.binding('natives')).filter(n => !/^_|\//.test(n));
 const nodeModules = ['electron', 'original-fs']
 	.concat(Object.keys(product.dependencies || {}))
@@ -43,15 +47,14 @@ const nodeModules = ['electron', 'original-fs']
 	.concat(baseModules);
 
 // Build
-
-const builtInExtensions = [
-	{ name: 'ms-vscode.node-debug', version: '1.20.0' },
-	{ name: 'ms-vscode.node-debug2', version: '1.19.4' }
-];
+// @ts-ignore Microsoft/TypeScript#21262 complains about a require of a JSON file
+const builtInExtensions = require('./builtInExtensions.json');
 
 const excludedExtensions = [
 	'vscode-api-tests',
-	'vscode-colorize-tests'
+	'vscode-colorize-tests',
+	'ms-vscode.node-debug',
+	'ms-vscode.node-debug2',
 ];
 
 const vscodeEntryPoints = _.flatten([
@@ -75,14 +78,15 @@ const vscodeResources = [
 	'out-build/vs/workbench/electron-browser/bootstrap/**',
 	'out-build/vs/workbench/parts/debug/**/*.json',
 	'out-build/vs/workbench/parts/execution/**/*.scpt',
-	'out-build/vs/workbench/parts/html/browser/webview-pre.js',
+	'out-build/vs/workbench/parts/html/electron-browser/webview-pre.js',
 	'out-build/vs/**/markdown.css',
 	'out-build/vs/workbench/parts/tasks/**/*.json',
 	'out-build/vs/workbench/parts/terminal/electron-browser/terminalProcess.js',
 	'out-build/vs/workbench/parts/welcome/walkThrough/**/*.md',
 	'out-build/vs/workbench/services/files/**/*.exe',
 	'out-build/vs/workbench/services/files/**/*.md',
-	'out-build/vs/code/electron-browser/sharedProcess.js',
+	'out-build/vs/code/electron-browser/sharedProcess/sharedProcess.js',
+	'out-build/vs/code/electron-browser/issue/issueReporter.js',
 	'!**/test/**'
 ];
 
@@ -92,10 +96,7 @@ const BUNDLED_FILE_HEADER = [
 	' *--------------------------------------------------------*/'
 ].join('\n');
 
-var languages = ['chs', 'cht', 'jpn', 'kor', 'deu', 'fra', 'esn', 'rus', 'ita'];
-if (process.env.VSCODE_QUALITY !== 'stable') {
-	languages = languages.concat(['ptb', 'hun', 'trk']); // Add languages requested by the community to non-stable builds
-}
+const languages = i18n.defaultLanguages.concat([]);  // i18n.defaultLanguages.concat(process.env.VSCODE_QUALITY !== 'stable' ? i18n.extraLanguages : []);
 
 gulp.task('clean-optimized-vscode', util.rimraf('out-vscode'));
 gulp.task('optimize-vscode', ['clean-optimized-vscode', 'compile-build', 'compile-extensions-build'], common.optimizeTask({
@@ -105,7 +106,8 @@ gulp.task('optimize-vscode', ['clean-optimized-vscode', 'compile-build', 'compil
 	loaderConfig: common.loaderConfig(nodeModules),
 	header: BUNDLED_FILE_HEADER,
 	out: 'out-vscode',
-	languages: languages
+	languages: languages,
+	bundleInfo: undefined
 }));
 
 
@@ -127,7 +129,7 @@ const config = {
 	version: getElectronVersion(),
 	productAppName: product.nameLong,
 	companyName: 'Microsoft Corporation',
-	copyright: 'Copyright (C) 2017 Microsoft. All rights reserved',
+	copyright: 'Copyright (C) 2018 Microsoft. All rights reserved',
 	darwinIcon: 'resources/darwin/code.icns',
 	darwinBundleIdentifier: product.darwinBundleIdentifier,
 	darwinApplicationCategoryType: 'public.app-category.developer-tools',
@@ -145,7 +147,7 @@ const config = {
 		name: product.nameLong,
 		urlSchemes: [product.urlProtocol]
 	}],
-	darwinCredits: darwinCreditsTemplate ? new Buffer(darwinCreditsTemplate({ commit: commit, date: new Date().toISOString() })) : void 0,
+	darwinCredits: darwinCreditsTemplate ? Buffer.from(darwinCreditsTemplate({ commit: commit, date: new Date().toISOString() })) : void 0,
 	linuxExecutableName: product.applicationName,
 	winIcon: 'resources/win32/code.ico',
 	token: process.env['VSCODE_MIXIN_PASSWORD'] || process.env['GITHUB_TOKEN'] || void 0,
@@ -304,9 +306,10 @@ function packageTask(platform, arch, opts) {
 			.pipe(util.cleanNodeModule('windows-process-tree', ['binding.gyp', 'build/**', 'src/**'], ['**/*.node']))
 			.pipe(util.cleanNodeModule('gc-signals', ['binding.gyp', 'build/**', 'src/**', 'deps/**'], ['**/*.node', 'src/index.js']))
 			.pipe(util.cleanNodeModule('keytar', ['binding.gyp', 'build/**', 'src/**', 'script/**', 'node_modules/**'], ['**/*.node']))
-			.pipe(util.cleanNodeModule('node-pty', ['binding.gyp', 'build/**', 'src/**', 'tools/**'], ['build/Release/**']))
-			.pipe(util.cleanNodeModule('nsfw', ['binding.gyp', 'build/**', 'src/**', 'openpa/**', 'includes/**'], ['**/*.node', '**/*.a']))
-			.pipe(util.cleanNodeModule('vsda', ['binding.gyp', 'README.md', 'build/**', '*.bat', '*.sh', '*.cpp', '*.h'], ['build/Release/vsda.node']));
+			.pipe(util.cleanNodeModule('node-pty', ['binding.gyp', 'build/**', 'src/**', 'tools/**'], ['build/Release/*.exe', 'build/Release/*.dll', 'build/Release/*.node']))
+			.pipe(util.cleanNodeModule('vscode-nsfw', ['binding.gyp', 'build/**', 'src/**', 'openpa/**', 'includes/**'], ['**/*.node', '**/*.a']))
+			.pipe(util.cleanNodeModule('vsda', ['binding.gyp', 'README.md', 'build/**', '*.bat', '*.sh', '*.cpp', '*.h'], ['build/Release/vsda.node']))
+			.pipe(createAsar(path.join(process.cwd(), 'node_modules'), ['**/*.node', '**/vscode-ripgrep/bin/*', '**/node-pty/build/Release/*'], 'app/node_modules.asar'));
 
 		let all = es.merge(
 			packageJsonStream,
@@ -382,25 +385,21 @@ gulp.task('vscode-linux-x64-min', ['minify-vscode', 'clean-vscode-linux-x64'], p
 gulp.task('vscode-linux-arm-min', ['minify-vscode', 'clean-vscode-linux-arm'], packageTask('linux', 'arm', { minified: true }));
 
 // Transifex Localizations
-const vscodeLanguages = [
-	'zh-hans',
-	'zh-hant',
-	'ja',
-	'ko',
-	'de',
-	'fr',
-	'es',
-	'ru',
-	'it',
-	'pt-br',
-	'hu',
-	'tr'
-];
-const setupDefaultLanguages = [
-	'zh-hans',
-	'zh-hant',
-	'ko'
-];
+
+const innoSetupConfig = {
+	'zh-cn': { codePage: 'CP936', defaultInfo: { name: 'Simplified Chinese', id: '$0804', } },
+	'zh-tw': { codePage: 'CP950', defaultInfo: { name: 'Traditional Chinese', id: '$0404' } },
+	'ko': { codePage: 'CP949', defaultInfo: { name: 'Korean', id: '$0412' } },
+	'ja': { codePage: 'CP932' },
+	'de': { codePage: 'CP1252' },
+	'fr': { codePage: 'CP1252' },
+	'es': { codePage: 'CP1252' },
+	'ru': { codePage: 'CP1251' },
+	'it': { codePage: 'CP1252' },
+	'pt-br': { codePage: 'CP1252' },
+	'hu': { codePage: 'CP1250' },
+	'tr': { codePage: 'CP1254' }
+};
 
 const apiHostname = process.env.TRANSIFEX_API_URL;
 const apiName = process.env.TRANSIFEX_API_NAME;
@@ -408,27 +407,48 @@ const apiToken = process.env.TRANSIFEX_API_TOKEN;
 
 gulp.task('vscode-translations-push', ['optimize-vscode'], function () {
 	const pathToMetadata = './out-vscode/nls.metadata.json';
-	const pathToExtensions = './extensions/**/*.nls.json';
+	const pathToExtensions = './extensions/*';
 	const pathToSetup = 'build/win32/**/{Default.isl,messages.en.isl}';
 
 	return es.merge(
-		gulp.src(pathToMetadata).pipe(i18n.prepareXlfFiles()),
-		gulp.src(pathToSetup).pipe(i18n.prepareXlfFiles()),
-		gulp.src(pathToExtensions).pipe(i18n.prepareXlfFiles('vscode-extensions'))
+		gulp.src(pathToMetadata).pipe(i18n.createXlfFilesForCoreBundle()),
+		gulp.src(pathToSetup).pipe(i18n.createXlfFilesForIsl()),
+		gulp.src(pathToExtensions).pipe(i18n.createXlfFilesForExtensions())
+	).pipe(i18n.findObsoleteResources(apiHostname, apiName, apiToken)
 	).pipe(i18n.pushXlfFiles(apiHostname, apiName, apiToken));
 });
 
-gulp.task('vscode-translations-pull', function () {
+gulp.task('vscode-translations-push-test', ['optimize-vscode'], function () {
+	const pathToMetadata = './out-vscode/nls.metadata.json';
+	const pathToExtensions = './extensions/*';
+	const pathToSetup = 'build/win32/**/{Default.isl,messages.en.isl}';
+
 	return es.merge(
-		i18n.pullXlfFiles('vscode-editor', apiHostname, apiName, apiToken, vscodeLanguages),
-		i18n.pullXlfFiles('vscode-workbench', apiHostname, apiName, apiToken, vscodeLanguages),
-		i18n.pullXlfFiles('vscode-extensions', apiHostname, apiName, apiToken, vscodeLanguages),
-		i18n.pullXlfFiles('vscode-setup', apiHostname, apiName, apiToken, setupDefaultLanguages)
-	).pipe(vfs.dest('../vscode-localization'));
+		gulp.src(pathToMetadata).pipe(i18n.createXlfFilesForCoreBundle()),
+		gulp.src(pathToSetup).pipe(i18n.createXlfFilesForIsl()),
+		gulp.src(pathToExtensions).pipe(i18n.createXlfFilesForExtensions())
+	).pipe(i18n.findObsoleteResources(apiHostname, apiName, apiToken)
+	).pipe(vfs.dest('../vscode-transifex-input'));
+});
+
+gulp.task('vscode-translations-pull', function () {
+	[...i18n.defaultLanguages, ...i18n.extraLanguages].forEach(language => {
+		i18n.pullCoreAndExtensionsXlfFiles(apiHostname, apiName, apiToken, language).pipe(vfs.dest(`../vscode-localization/${language.id}/build`));
+
+		let includeDefault = !!innoSetupConfig[language.id].defaultInfo;
+		i18n.pullSetupXlfFiles(apiHostname, apiName, apiToken, language, includeDefault).pipe(vfs.dest(`../vscode-localization/${language.id}/setup`));
+	});
 });
 
 gulp.task('vscode-translations-import', function () {
-	return gulp.src('../vscode-localization/**/*.xlf').pipe(i18n.prepareJsonFiles()).pipe(vfs.dest('./i18n'));
+	[...i18n.defaultLanguages, ...i18n.extraLanguages].forEach(language => {
+		gulp.src(`../vscode-localization/${language.id}/build/*/*.xlf`)
+			.pipe(i18n.prepareI18nFiles())
+			.pipe(vfs.dest(`./i18n/${language.folderName}`));
+		gulp.src(`../vscode-localization/${language.id}/setup/*/*.xlf`)
+			.pipe(i18n.prepareIslFiles(language, innoSetupConfig[language.id]))
+			.pipe(vfs.dest(`./build/win32/i18n`));
+	});
 });
 
 // Sourcemaps
@@ -454,7 +474,8 @@ gulp.task('upload-vscode-sourcemaps', ['minify-vscode'], () => {
 const allConfigDetailsPath = path.join(os.tmpdir(), 'configuration.json');
 gulp.task('upload-vscode-configuration', ['generate-vscode-configuration'], () => {
 	const branch = process.env.BUILD_SOURCEBRANCH;
-	if (!branch.endsWith('/master') && branch.indexOf('/release/') < 0) {
+
+	if (!/\/master$/.test(branch) && branch.indexOf('/release/') < 0) {
 		console.log(`Only runs on master and release branches, not ${branch}`);
 		return;
 	}
@@ -549,6 +570,7 @@ function versionStringToNumber(versionStr) {
 	return parseInt(match[1], 10) * 1e4 + parseInt(match[2], 10) * 1e2 + parseInt(match[3], 10);
 }
 
+// This task is only run for the MacOS build
 gulp.task('generate-vscode-configuration', () => {
 	return new Promise((resolve, reject) => {
 		const buildDir = process.env['AGENT_BUILDDIRECTORY'];
@@ -558,7 +580,8 @@ gulp.task('generate-vscode-configuration', () => {
 
 		const userDataDir = path.join(os.tmpdir(), 'tmpuserdata');
 		const extensionsDir = path.join(os.tmpdir(), 'tmpextdir');
-		const appPath = path.join(buildDir, 'VSCode-darwin/Visual\\ Studio\\ Code\\ -\\ Insiders.app/Contents/Resources/app/bin/code');
+		const appName = process.env.VSCODE_QUALITY === 'insider' ? 'Visual\\ Studio\\ Code\\ -\\ Insiders.app' : 'Visual\\ Studio\\ Code.app';
+		const appPath = path.join(buildDir, `VSCode-darwin/${appName}/Contents/Resources/app/bin/code`);
 		const codeProc = cp.exec(`${appPath} --export-default-configuration='${allConfigDetailsPath}' --wait --user-data-dir='${userDataDir}' --extensions-dir='${extensionsDir}'`);
 
 		const timer = setTimeout(() => {
