@@ -20,6 +20,7 @@ const sourcemaps = require('gulp-sourcemaps');
 const nlsDev = require('vscode-nls-dev');
 const root = path.dirname(__dirname);
 const commit = util.getVersion(root);
+const i18n = require('./lib/i18n');
 
 const extensionsPath = path.join(path.dirname(__dirname), 'extensions');
 
@@ -29,7 +30,8 @@ const compilations = glob.sync('**/tsconfig.json', {
 });
 
 const getBaseUrl = out => `https://ticino.blob.core.windows.net/sourcemaps/${commit}/${out}`;
-const languages = ['chs', 'cht', 'jpn', 'kor', 'deu', 'fra', 'esn', 'rus', 'ita'];
+
+const languages = i18n.defaultLanguages.concat(process.env.VSCODE_QUALITY !== 'stable' ? i18n.extraLanguages : []);
 
 const tasks = compilations.map(function (tsconfigFile) {
 	const absolutePath = path.join(extensionsPath, tsconfigFile);
@@ -55,8 +57,18 @@ const tasks = compilations.map(function (tsconfigFile) {
 	const srcBase = path.join(root, 'src');
 	const src = path.join(srcBase, '**');
 	const out = path.join(root, 'out');
-	const i18n = path.join(__dirname, '..', 'i18n');
+	const i18nPath = path.join(__dirname, '..', 'i18n');
 	const baseUrl = getBaseUrl(out);
+
+	let headerId, headerOut;
+	let index = relativeDirname.indexOf('/');
+	if (index < 0) {
+		headerId = 'vscode.' + relativeDirname;
+		headerOut = 'out';
+	} else {
+		headerId = 'vscode.' + relativeDirname.substr(0, index);
+		headerOut = relativeDirname.substr(index + 1) + '/out';
+	}
 
 	function createPipeline(build, emitError) {
 		const reporter = createReporter();
@@ -82,7 +94,9 @@ const tasks = compilations.map(function (tsconfigFile) {
 					sourceRoot: '../src'
 				}))
 				.pipe(tsFilter.restore)
-				.pipe(build ? nlsDev.createAdditionalLanguageFiles(languages, i18n, out) : es.through())
+				.pipe(build ? nlsDev.createAdditionalLanguageFiles(languages, i18nPath, out) : es.through())
+				.pipe(build ? nlsDev.bundleMetaDataFiles(headerId, headerOut) : es.through())
+				.pipe(build ? nlsDev.bundleLanguageFiles() : es.through())
 				.pipe(reporter.end(emitError));
 
 			return es.duplex(input, output);
@@ -129,7 +143,7 @@ const tasks = compilations.map(function (tsconfigFile) {
 		const watchInput = watcher(src, srcOpts);
 
 		return watchInput
-			.pipe(util.incremental(() => pipeline(true), input))
+			.pipe(util.incremental(() => pipeline(), input))
 			.pipe(gulp.dest(out));
 	});
 
