@@ -12,34 +12,14 @@ const shell = require('gulp-shell');
 const es = require('event-stream');
 const vfs = require('vinyl-fs');
 const util = require('./lib/util');
+// @ts-ignore Microsoft/TypeScript#21262 complains about a require of a JSON file
 const packageJson = require('../package.json');
+// @ts-ignore Microsoft/TypeScript#21262 complains about a require of a JSON file
 const product = require('../product.json');
-const rpmDependencies = require('../resources/linux/rpm/dependencies');
+// @ts-ignore Microsoft/TypeScript#21262 complains about a require of a JSON file
+const rpmDependencies = require('../resources/linux/rpm/dependencies.json');
 
 const linuxPackageRevision = Math.floor(new Date().getTime() / 1000);
-
-const flatpakManifest = {
-	appId: product.darwinBundleIdentifier,  // We need a reverse-url style identifier.
-	sdk: 'org.freedesktop.Sdk',
-	runtime: 'org.freedesktop.Sdk',
-	runtimeVersion: '1.4',
-	base: 'io.atom.electron.BaseApp',
-	baseFlatpakref: 'https://s3-us-west-2.amazonaws.com/electron-flatpak.endlessm.com/electron-base-app-master.flatpakref',
-	command: product.applicationName,
-	symlinks: [
-		['/share/' + product.applicationName + '/bin/' + product.applicationName, '/bin/' + product.applicationName],
-	],
-	finishArgs: [
-		'--share=ipc', '--socket=x11',  // Allow showing X11 windows.
-		'--share=network',              // Network access (e.g. for installing extension).
-		'--filesystem=host',            // Allow access to the whole file system.
-		'--device=dri',                 // Allow OpenGL rendering.
-		'--filesystem=/tmp',            // Needed for Chromium's single instance check.
-		'--socket=pulseaudio',          // Some extensions may want to play sounds...
-		'--talk-name=org.freedesktop.Notifications',  // ...or pop up notifications.
-	],
-};
-
 
 function getDebPackageArch(arch) {
 	return { x64: 'amd64', ia32: 'i386', arm: 'armhf' }[arch];
@@ -110,8 +90,7 @@ function buildDebPackage(arch) {
 	return shell.task([
 		'chmod 755 ' + product.applicationName + '-' + debArch + '/DEBIAN/postinst ' + product.applicationName + '-' + debArch + '/DEBIAN/prerm ' + product.applicationName + '-' + debArch + '/DEBIAN/postrm',
 		'mkdir -p deb',
-		'fakeroot dpkg-deb -b ' + product.applicationName + '-' + debArch + ' deb',
-		'dpkg-scanpackages deb /dev/null > Packages'
+		'fakeroot dpkg-deb -b ' + product.applicationName + '-' + debArch + ' deb'
 	], { cwd: '.build/linux/deb/' + debArch });
 }
 
@@ -218,84 +197,11 @@ function prepareSnapPackage(arch) {
 
 function buildSnapPackage(arch) {
 	const snapBuildPath = getSnapBuildPath(arch);
-
+	const snapFilename = `${product.applicationName}-${packageJson.version}-${linuxPackageRevision}-${arch}.snap`;
 	return shell.task([
 		`chmod +x ${snapBuildPath}/electron-launch`,
-		`cd ${snapBuildPath} && snapcraft snap`
+		`cd ${snapBuildPath} && snapcraft snap --output ../${snapFilename}`
 	]);
-}
-
-function getFlatpakArch(arch) {
-	return { x64: 'x86_64', ia32: 'i386', arm: 'arm' }[arch];
-}
-
-function prepareFlatpak(arch) {
-	const binaryDir = '../VSCode-linux-' + arch;
-	const flatpakArch = getFlatpakArch(arch);
-	const destination = '.build/linux/flatpak/' + flatpakArch;
-
-	return function () {
-		// This is not imported in the global scope to avoid requiring ImageMagick
-		// (or GraphicsMagick) when not building building Flatpak bundles.
-		const imgResize = require('gulp-image-resize');
-
-		const all = [16, 24, 32, 48, 64, 128, 192, 256, 512].map(function (size) {
-			return gulp.src('resources/linux/code.png', { base: '.' })
-				.pipe(imgResize({ width: size, height: size, format: "png", noProfile: true }))
-				.pipe(rename('share/icons/hicolor/' + size + 'x' + size + '/apps/' + flatpakManifest.appId + '.png'));
-		});
-
-		all.push(gulp.src('resources/linux/code.desktop', { base: '.' })
-			.pipe(replace('Exec=/usr/share/@@NAME@@/@@NAME@@', 'Exec=' + product.applicationName))
-			.pipe(replace('@@NAME_LONG@@', product.nameLong))
-			.pipe(replace('@@NAME_SHORT@@', product.nameShort))
-			.pipe(replace('@@NAME@@', product.applicationName))
-			.pipe(rename('share/applications/' + flatpakManifest.appId + '.desktop')));
-
-		all.push(gulp.src('resources/linux/code.appdata.xml', { base: '.' })
-			.pipe(replace('@@NAME_LONG@@', product.nameLong))
-			.pipe(replace('@@NAME@@', flatpakManifest.appId))
-			.pipe(replace('@@LICENSE@@', product.licenseName))
-			.pipe(rename('share/appdata/' + flatpakManifest.appId + '.appdata.xml')));
-
-		all.push(gulp.src(binaryDir + '/**/*', { base: binaryDir })
-			.pipe(rename(function (p) {
-				p.dirname = 'share/' + product.applicationName + '/' + p.dirname;
-			})));
-
-		return es.merge(all).pipe(vfs.dest(destination));
-	};
-}
-
-function buildFlatpak(arch) {
-	const flatpakArch = getFlatpakArch(arch);
-	const manifest = {};
-	for (var k in flatpakManifest) {
-		manifest[k] = flatpakManifest[k];
-	}
-	manifest.files = [
-		['.build/linux/flatpak/' + flatpakArch, '/'],
-	];
-	const buildOptions = {
-		arch: flatpakArch,
-		subject: product.nameLong + ' ' + packageJson.version + '.' + linuxPackageRevision,
-	};
-	// If requested, use the configured path for the OSTree repository.
-	if (process.env.FLATPAK_REPO) {
-		buildOptions.repoDir = process.env.FLATPAK_REPO;
-	} else {
-		buildOptions.bundlePath = manifest.appId + '-' + flatpakArch + '.flatpak';
-	}
-	// Setup PGP signing if requested.
-	if (process.env.GPG_KEY_ID !== undefined) {
-		buildOptions.gpgSign = process.env.GPG_KEY_ID;
-		if (process.env.GPG_HOMEDIR) {
-			buildOptions.gpgHomedir = process.env.GPG_HOME_DIR;
-		}
-	}
-	return function (cb) {
-		require('flatpak-bundler').bundle(manifest, buildOptions, cb);
-	};
 }
 
 gulp.task('clean-vscode-linux-ia32-deb', util.rimraf('.build/linux/deb/i386'));
@@ -331,10 +237,3 @@ gulp.task('vscode-linux-arm-prepare-snap', ['clean-vscode-linux-arm-snap'], prep
 gulp.task('vscode-linux-ia32-build-snap', ['vscode-linux-ia32-prepare-snap'], buildSnapPackage('ia32'));
 gulp.task('vscode-linux-x64-build-snap', ['vscode-linux-x64-prepare-snap'], buildSnapPackage('x64'));
 gulp.task('vscode-linux-arm-build-snap', ['vscode-linux-arm-prepare-snap'], buildSnapPackage('arm'));
-
-gulp.task('vscode-linux-ia32-prepare-flatpak', ['clean-vscode-linux-ia32-flatpak'], prepareFlatpak('ia32'));
-gulp.task('vscode-linux-x64-prepare-flatpak', ['clean-vscode-linux-x64-flatpak'], prepareFlatpak('x64'));
-gulp.task('vscode-linux-arm-prepare-flatpak', ['clean-vscode-linux-arm-flatpak'], prepareFlatpak('arm'));
-gulp.task('vscode-linux-ia32-flatpak', ['vscode-linux-ia32-prepare-flatpak'], buildFlatpak('ia32'));
-gulp.task('vscode-linux-x64-flatpak', ['vscode-linux-x64-prepare-flatpak'], buildFlatpak('x64'));
-gulp.task('vscode-linux-arm-flatpak', ['vscode-linux-arm-prepare-flatpak'], buildFlatpak('arm'));
