@@ -24,7 +24,8 @@ function watch(root) {
 	var result = es.through();
 	var child = cp.spawn(watcherPath, [root]);
 
-	child.stdout.on('data', function(data) {
+	child.stdout.on('data', function (data) {
+		// @ts-ignore
 		var lines = data.toString('utf8').split('\n');
 		for (var i = 0; i < lines.length; i++) {
 			var line = lines[i].trim();
@@ -46,17 +47,17 @@ function watch(root) {
 				path: changePathFull,
 				base: root
 			});
-
+			//@ts-ignore
 			file.event = toChangeType(changeType);
 			result.emit('data', file);
 		}
 	});
 
-	child.stderr.on('data', function(data) {
+	child.stderr.on('data', function (data) {
 		result.emit('error', data);
 	});
 
-	child.on('exit', function(code) {
+	child.on('exit', function (code) {
 		result.emit('error', 'Watcher died with code ' + code);
 		child = null;
 	});
@@ -70,7 +71,7 @@ function watch(root) {
 
 var cache = Object.create(null);
 
-module.exports = function(pattern, options) {
+module.exports = function (pattern, options) {
 	options = options || {};
 
 	var cwd = path.normalize(options.cwd || process.cwd());
