@@ -22,33 +22,22 @@ function yarnInstall(location, opts) {
 
 yarnInstall('extensions'); // node modules shared by all extensions
 
-const extensions = [
-	'vscode-api-tests',
-	'vscode-colorize-tests',
-	'json',
-	'configuration-editing',
-	'extension-editing',
-	'markdown',
-	'typescript',
-	'php',
-	'javascript',
-	'css',
-	'html',
-	'git',
-	'gulp',
-	'grunt',
-	'jake',
-	'merge-conflict',
-	'emmet',
-	'npm',
-	'jake'
-];
+const allExtensionFolders = fs.readdirSync('extensions');
+const extensions = allExtensionFolders.filter(e => {
+	try {
+		let packageJSON = JSON.parse(fs.readFileSync(path.join('extensions', e, 'package.json')).toString());
+		return packageJSON && (packageJSON.dependencies || packageJSON.devDependencies);
+	} catch (e) {
+		return false;
+	}
+});
 
 extensions.forEach(extension => yarnInstall(`extensions/${extension}`));
 
 function yarnInstallBuildDependencies() {
 	// make sure we install the deps of build/lib/watch for the system installed
 	// node, since that is the driver of gulp
+	//@ts-ignore
 	const env = Object.assign({}, process.env);
 	const watchPath = path.join(path.dirname(__dirname), 'lib', 'watch');
 	const yarnrcPath = path.join(watchPath, '.yarnrc');
